# openapi.yaml
openapi: 3.1.0
info:
  title: Audicle
  version: 0.57.0
paths:
...
  /media/{episode_id}.mp3:
    head:
      tags:
      - media
      summary: Get Mp3
      operationId: head_media_mp3
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
    get:
      tags:
      - media
      summary: Get Mp3
      operationId: get_media_mp3
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
  /media/{episode_id}.jpg:
    head:
      tags:
      - media
      summary: Get Jpg
      operationId: head_media_jpg
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
    get:
      tags:
      - media
      summary: Get Jpg
      operationId: get_media_jpg
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
  /media/{episode_id}.vtt:
    head:
      tags:
      - media
      summary: Get Vtt
      operationId: head_media_vtt
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
    get:
      tags:
      - media
      summary: Get Vtt
      operationId: get_media_vtt
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
  /media/{episode_id}.chapters.json:
    head:
      tags:
      - media
      summary: Get Chapters
      description: 'The Podcasting 2.0 chapters document, served from the ``chapters_json``

        column. 404 for episodes without chapters (short, disabled, or pre-0.51.0).'
      operationId: head_media_chapters
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
    get:
      tags:
      - media
      summary: Get Chapters
      description: 'The Podcasting 2.0 chapters document, served from the ``chapters_json``

        column. 404 for episodes without chapters (short, disabled, or pre-0.51.0).'
      operationId: get_media_chapters
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
  /media/{episode_id}.txt:
    head:
      tags:
      - media
      summary: Get Cleaned Text
      description: 'The cleaned article text (the exact input to TTS), served from
        the

        ``cleaned_text`` column. 404 for episodes processed before 0.6.0 (NULL).'
      operationId: head_media_text
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
    get:
      tags:
      - media
      summary: Get Cleaned Text
      description: 'The cleaned article text (the exact input to TTS), served from
        the

        ``cleaned_text`` column. 404 for episodes processed before 0.6.0 (NULL).'
      operationId: get_media_text
      parameters:
      - name: episode_id
        in: path
        required: true
        schema:
          type: string
          title: Episode Id
      - name: v
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: V
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - FeedKey: []
      - AdminSession: []
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - status
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: audicle_session
      description: Required for the admin API when a password is configured.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the admin session for unsafe methods.
    FeedKey:
      type: apiKey
      in: query
      name: key
      description: Required for protected feeds and non-artwork media.